proactive_sentinel_agent/requirements.txt
google-adk>=0.1.0
google-cloud-aiplatform>=1.38.0
opentelemetry-api
opentelemetry-sdk
python-dotenv
pandas
pytest
uvicorn
fastapi